split functionality into smaller components
