+++ tests/components/about/Employment.test.jsx
@@ -3,6 +3,7 @@ import {render} from '@testing-library/react';
 import Chance from 'chance';
 
 import {Employment} from '../../../src/components/about/Employment';
+import { createExternalProject, createInternalProject } from '../../utils/employmentData';
 
 const chance = new Chance();
 
@@ -76,20 +77,8 @@ describe('Employment', () => {
                 Employer: chance.company(),
                 End_Date: chance.date({string: true}),
                 Projects: {
-                    External: [{
-                        Description: chance.n(chance.sentence, chance.d4()),
-                        End_Date: chance.date({string: true}),
-                        Partner_Site: chance.company(),
-                        Start_Date: chance.date({string: true}),
-                        Technology: chance.n(chance.word, chance.d4())
-                    }],
-                    Internal: [{
-                        Description: chance.n(chance.sentence, chance.d4()),
-                        End_Date: chance.date({string: true}),
-                        Name: chance.word(),
-                        Start_Date: chance.date({string: true}),
-                        Technology: chance.n(chance.word, chance.d4())
-                    }]                    
+                    External: [createExternalProject()],
+                    Internal: [createInternalProject()]                    
                 },
                 Start_Date: chance.date({string: true}),
                 Title: chance.profession()                
@@ -128,34 +117,6 @@ describe('Employment', () => {
                 expect(internalHeader).toBeDefined();
             });
 
-            it('should display the Project Name', () => {
-                const projectName = testRenderResult.getByText(employmentData.Projects.Internal[0].Name);
-    
-                expect(projectName).toBeDefined();
-            });
-
-            it('should display the Dates spent on the Project', () => {
-                const startDate = employmentData.Projects.Internal[0].Start_Date;
-                const endDate = employmentData.Projects.Internal[0].End_Date;
-                const datesWOrked = testRenderResult.getByText(`From ${startDate} - ${endDate}`);
-    
-                expect(datesWOrked).toBeDefined();
-            });
-
-            it('should display the Project Description', () => {
-                const description = employmentData.Projects.Internal[0].Description[0];                
-                const projectDescription = testRenderResult.getByText(description);
-    
-                expect(projectDescription).toBeDefined();
-            });
-
-            it('should display the Technology used on the Project', () => {
-                const technology = employmentData.Projects.Internal[0].Technology[0];
-                const projectTech = testRenderResult.getByText(technology);
-    
-                expect(projectTech).toBeDefined();
-            });
-
             it('should **not** display the Internal Projects if there are none', () => {
                 employmentData.Projects.Internal = null;
                 testRenderResult.rerender(<Employment {...employmentData} />);
@@ -173,37 +134,9 @@ describe('Employment', () => {
                 expect(externalHeader).toBeDefined();
             });
 
-            it('should display the Partner Site Name', () => {
-                const partnerName = employmentData.Projects.External[0].Partner_Site;
-                const partnerSite = testRenderResult.getByText(partnerName);
-
-                expect(partnerSite).toBeDefined();
-            });
-
-            it('should display the Dates worked for the Partner Site', () => {
-                const startDate = employmentData.Projects.External[0].Start_Date;
-                const endDate = employmentData.Projects.External[0].End_Date;
-                const datesWorked = testRenderResult.getByText(`From ${startDate} - ${endDate}`);
-
-                expect(datesWorked).toBeDefined();
-            });
-
-            it('should display the Description from the Partner Site', () => {
-                const partnerDescription = employmentData.Projects.External[0].Description[0];
-                const description = testRenderResult.getByText(partnerDescription);
-
-                expect(description).toBeDefined();
-            });
-
-            it('should display the Technology used at the Partner Site', () => {
-                const tech = employmentData.Projects.External[0].Technology[0];
-                const techUsed = testRenderResult.getByText(tech);
-
-                expect(techUsed).toBeDefined();
-            });
-
             it('should **not** display the External Projects header if there are none', () => {
                 employmentData.Projects.External = null;
+
                 testRenderResult.rerender(<Employment {...employmentData} />);
 
                 const externalHeader = testRenderResult.queryByText('External Projects');
